删除不需使用的模型与字段
